Moved InProcessAdapter and ClientGalaxy to the API project,
Added a `ClientFactory` static class to provide utility methods that create clients.
